% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{addGeneAnnotation}
\alias{addGeneAnnotation}
\title{Add geneannotation to one read}
\usage{
addGeneAnnotation(readSeqnames, readStart, readWidth, dataType = "hg38")
}
\arguments{
\item{readSeqnames}{An integer representing readSeqnames}

\item{readStart}{A position showing read start}

\item{readWidth}{An integer showing read width}

\item{dataType}{datatype relating to a database}
}
\value{
A list with possible read pairs
}
\description{
Add geneannotation to one read
}
