% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{addHeat}
\alias{addHeat}
\title{Add heat to the line we want}
\usage{
addHeat(formattedBedfile, list, heatnumber)
}
\arguments{
\item{formattedBedfile}{A bed file}

\item{list}{A list included name of reads we want to add heat}

\item{heatnumber}{number you want to added}
}
\value{
A bed file filled with base col in each line
}
\description{
Add heat to the line we want
}
