% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{changeFormatChromoMap}
\alias{changeFormatChromoMap}
\title{Change format for visualization by chromoMap}
\usage{
changeFormatChromoMap(bedfile, baseCol)
}
\arguments{
\item{bedfile}{A bed file}

\item{baseCol}{basecolour parameter}
}
\value{
A bed file filled with base col in each line
}
\description{
Change format for visualization by chromoMap
}
