% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{checkDistanceAllReadPair}
\alias{checkDistanceAllReadPair}
\title{Check distance requirment on all read pair appeared in prePair list}
\usage{
checkDistanceAllReadPair(prePair, maxDistance)
}
\arguments{
\item{prePair}{read pairs met before quirments}

\item{maxDistance}{The distance requirment}
}
\value{
A list with possible read pairs
}
\description{
Check distance requirment on all read pair appeared in prePair list
}
\examples{
dupFiltedBedFile2 <- readRDS(system.file('extdata', 'dupFiltedBedFile2.Rda', package = 'BreakViz'))
prePair  <- combineSameRead(bedFile2, dupFiltedBedFile2)
checkDistanceAllReadPair(prePair, maxDistance)
}
