% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{checkDistanceSingleReadPair}
\alias{checkDistanceSingleReadPair}
\title{Check distance between each parts in one read pair if met the distance requirement}
\usage{
checkDistanceSingleReadPair(maxDistance, read1part1Start, read1part1Range,
  read1part2Start, read1part2Range, read2part1Start, read2part1Range,
  read2part2Start, read2part2Range)
}
\arguments{
\item{maxDistance}{The distance requirment}

\item{read1part1Start}{The start position of part1 of read1}

\item{read1part2Start}{The start position of part2 of read1}

\item{read2part1Start}{The start position of part1 of read2}

\item{read2part2Start}{The start position of part2 of read2}

\item{read1part1SRange}{The range of part1 of read1}

\item{read1part2SRange}{The range of part2 of read1}

\item{read2part1SRange}{The range of part1 of read2}

\item{read2part2SRange}{The range of part2 of read2}
}
\value{
The logical value showing if there is a read pair met distance requirment
}
\description{
Check distance between each parts in one read pair if met the distance requirement
}
\examples{
bedFile2 <- rtracklayer::import(system.file('extdata', 'test_file_2.bed', package = 'BreakViz'), format = 'bed')

}
