% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{combineSameRead}
\alias{combineSameRead}
\title{Combine reads that only mapping to two parts of chromosome and two parts are unique to each other}
\usage{
combineSameRead(bedfile, preFiltedRead)
}
\arguments{
\item{bedfile}{A bed file after sizeFilter}

\item{preFiltedRead}{A file after sizeFilter and dupFilter}
}
\value{
the list of read that only mapping to two parts of chromosome and their information of each part is included
}
\description{
Combine reads that only mapping to two parts of chromosome and two parts are unique to each other
}
