% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{copyInfo}
\alias{copyInfo}
\title{Copy information from outFile to inFile}
\usage{
copyInfo(inFile, numRowInFile, outFile, numRowOutFile, numCol)
}
\arguments{
\item{inFile}{File we want to transfer into}

\item{numRowInFile}{The number of rows in inFile we want to transfer information}

\item{outFile}{File we want to transfer from}

\item{numRowOutFile}{The number of rows in inFile we want to transfer information into}
}
\description{
Copy information from outFile to inFile
}
