% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{dupFileter}
\alias{dupFileter}
\title{Select reads that only mapping to two parts of chromosome}
\usage{
dupFileter(bedFile)
}
\arguments{
\item{bedFile}{A bed file after sizeFilter}
}
\value{
the list of read name that only mapping to two parts of chromosome
}
\description{
Select reads that only mapping to two parts of chromosome
}
\examples{
bedFile1 <- rtracklayer::import(system.file('extdata', 'test_file_1.bed', package = 'BreakViz'))
dupFileter(bedFile1)
}
