% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{getGeneAnnotation}
\alias{getGeneAnnotation}
\title{Add geneannotation to each read in a file}
\usage{
getGeneAnnotation(breakPair, dataType = "hg38")
}
\arguments{
\item{breakPair}{An file including read pairs}

\item{dataType}{datatype relating to a database}
}
\value{
A list with read pairs that have been added gene annotation
}
\description{
Add geneannotation to each read in a file
}
