% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{searchPossiblePairs}
\alias{searchPossiblePairs}
\title{Search possible pairs}
\usage{
searchPossiblePairs(bedfile, minOverlap = 1000, maxDistance = 100)
}
\arguments{
\item{bedfile}{A bed file}

\item{minOverlap}{A sizeOverlapping data}

\item{maxDistance}{The distance requirment}
}
\value{
A list with possible read pairs
}
\description{
Search possible pairs
}
