% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{sizefilter}
\alias{sizefilter}
\title{Filter reads with size requirement}
\usage{
sizefilter(bedFile, minOverlap = 1000)
}
\arguments{
\item{bedFile}{A bed file}

\item{minOverlap}{A length that a part of one read aligned to chromosomes}
}
\value{
the list of read parts met our size requirment.
}
\description{
Filter reads with size requirement
}
\examples{
bedFile1 <- rtracklayer::import(system.file('extdata', 'test_file_1.bed', package = 'BreakViz'))
sizefilter(bedFile1,10)
}
