% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{visPossiblePair}
\alias{visPossiblePair}
\title{Visualize possible pairs}
\usage{
visPossiblePair(bedfile, minOverlap = 1000, maxDistance = 100,
  baseCol = 1)
}
\arguments{
\item{bedfile}{A bed file}

\item{minOverlap}{A sizeOverlapping data}

\item{maxDistance}{The distance requirment}

\item{baseCol}{The base color}
}
\value{
A list with possible read pairs
}
\description{
Visualize possible pairs
}
