% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_viz.R
\name{changeFormatChromoMap}
\alias{changeFormatChromoMap}
\title{Change format for visualization by chromoMap}
\usage{
changeFormatChromoMap(bedfile)
}
\arguments{
\item{bedfile}{A bed file}
}
\value{
A bed file filled with base col in each line
}
\description{
Change format for visualization by chromoMap
}
\examples{
bedFile <- import(system.file('extdata', 'test_file_4.bed', package = 'BreakViz'), format = 'bed')
changeFormatChromoMap(bedFile)
}
