% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{combineSameRead}
\alias{combineSameRead}
\title{Combine reads that only mapping to two parts of chromosome and two parts are unique to each other}
\usage{
combineSameRead(bedFile, preFiltedRead)
}
\arguments{
\item{bedFile}{A bed file after sizeFilter}

\item{preFiltedRead}{A file after sizeFilter and dupFilter}
}
\value{
the list of read that only mapping to two parts of chromosome and their information of each part is included
}
\description{
Combine reads that only mapping to two parts of chromosome and two parts are unique to each other
}
\examples{
bedFile2 <- rtracklayer::import(system.file('extdata', 'test_file_2.bed', package = 'BreakViz'), format = 'bed')
dupFiltedBedFile2 <- readRDS(system.file('extdata', 'dupFiltedBedFile2.Rda', package = 'BreakViz'))
combineSameRead(bedFile2, dupFiltedBedFile2)
}
