% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{dupFileter}
\alias{dupFileter}
\title{Selet reads that only mapping to two parts of chromosome}
\usage{
dupFileter(file)
}
\arguments{
\item{bedfile}{A bed file after sizeFilter}
}
\value{
the list of read name that only mapping to two parts of chromosome
}
\description{
Selet reads that only mapping to two parts of chromosome
}
