% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{sizefilter}
\alias{sizefilter}
\title{Filt reads with size requirment}
\usage{
sizefilter(bedfile, minOverlap = 1000)
}
\arguments{
\item{bedfile}{A bed file}

\item{minOverlap}{A sizeOverlapping data}
}
\value{
the list of read parts met our size requirment.
}
\description{
Filt reads with size requirment
}
