% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreak.R
\name{sortChrom}
\alias{sortChrom}
\title{Sort reads parts according to their chromosome number}
\usage{
sortChrom(lists)
}
\arguments{
\item{lists}{A line including read part information of one read with chromosome number}
}
\value{
A line whose chromosome number has been ordered.
}
\description{
Sort reads parts according to their chromosome number
}
