% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz.R
\name{visPossiblePair}
\alias{visPossiblePair}
\title{Visualize possible pairs}
\usage{
visPossiblePair(bedfile, minOverlap = 100, maxDistance = 1000,
  colorSet = 1)
}
\arguments{
\item{bedfile}{A bed file that contains possible pairs for visualization.}

\item{minOverlap}{A length that a part of one read aligned to chromosomes}

\item{maxDistance}{A distance that the parts of one pair between each other}
}
\value{
A list with possible read pairs
}
\description{
Visualize possible pairs
}
\examples{
bedFile <- import(system.file('extdata', 'test_file_4.bed', package = 'BreakViz'), format = 'bed')
visPossiblePair(bedFile, minOverlap = 100, maxDistance = 1000, 1)
}
